% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{analytical_age_of_lx}
\alias{analytical_age_of_lx}
\title{Get age at which X prop of survival remains}
\usage{
analytical_age_of_lx(a, b, X = 0.5)
}
\description{
Get age at which X prop of survival remains
}
