% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{combmodplot_ageXaccuracy}
\alias{combmodplot_ageXaccuracy}
\title{Combined age X plot}
\usage{
combmodplot_ageXaccuracy(
  X,
  post,
  age.seq,
  species.is,
  minages = NULL,
  names.key = NULL,
  return.data = FALSE
)
}
\description{
Combined age X plot
}
