% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{combmodplot_curvecomp}
\alias{combmodplot_curvecomp}
\title{Latent Model predicted survival curves}
\usage{
combmodplot_curvecomp(
  post,
  age.seq,
  input.list,
  names.key = NULL,
  sim.in = NULL,
  N = 150
)
}
\description{
Latent Model predicted survival curves
}
