% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodcode_function.R
\name{get_marinesurvival_manyspecies.oneormanydatasets_stancode}
\alias{get_marinesurvival_manyspecies.oneormanydatasets_stancode}
\title{Generate a text file of the stan mode to be used for the latent model}
\usage{
get_marinesurvival_manyspecies.oneormanydatasets_stancode()
}
\description{
Generate a text file of the stan mode to be used for the latent model
}
