% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodcode_function preR3.R,
%   R/stanmodcode_function.R
\name{get_marinesurvival_onespecies.manydatasets_stancode}
\alias{get_marinesurvival_onespecies.manydatasets_stancode}
\title{Generate a text file of the stan mode to be used for the latent model adapted for if there is only one species}
\usage{
get_marinesurvival_onespecies.manydatasets_stancode()

get_marinesurvival_onespecies.manydatasets_stancode()
}
\description{
Generate a text file of the stan mode to be used for the latent model adapted for if there is only one species

Generate a text file of the stan mode to be used for the latent model adapted for if there is only one species
}
