% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodcode_function.R
\name{get_marinesurvival_onespecies.onedataset_stancode}
\alias{get_marinesurvival_onespecies.onedataset_stancode}
\title{Generate a text file of the stan mode to be used}
\usage{
get_marinesurvival_onespecies.onedataset_stancode()
}
\description{
Generate a text file of the stan mode to be used
}
