% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanmodcode_function.R
\name{get_marinesurvival_stancode}
\alias{get_marinesurvival_stancode}
\title{Choose model and generate approiate code}
\usage{
get_marinesurvival_stancode(mod.data.list)
}
\description{
Choose model and generate approiate code
}
