% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{latentmodplot_ageXaccuracy}
\alias{latentmodplot_ageXaccuracy}
\title{Prediction Survial to Age plot latent model version}
\usage{
latentmodplot_ageXaccuracy(
  X,
  post,
  age.seq,
  species.is,
  minages = NULL,
  names.key = NULL,
  return.data = FALSE
)
}
\description{
Prediction Survial to Age plot latent model version
}
