% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{latentmodplot_curvecomp}
\alias{latentmodplot_curvecomp}
\title{Latent Model predicted survival curves}
\usage{
latentmodplot_curvecomp(post, age.seq, input.list, names.key = NULL)
}
\description{
Latent Model predicted survival curves
}
