% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{modplot_ageXaccuracy}
\alias{modplot_ageXaccuracy}
\title{Prediction Survial to Age plot}
\usage{
modplot_ageXaccuracy(X, post, age.seq, sim.in = NULL)
}
\description{
Prediction Survial to Age plot
}
