% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{modplot_curvecomp}
\alias{modplot_curvecomp}
\title{Model predicted survival curves}
\usage{
modplot_curvecomp(sim.in = NULL, post, age.seq)
}
\description{
Model predicted survival curves
}
