% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{modplot_samplescheck}
\alias{modplot_samplescheck}
\title{Model to Sample Comparison plot}
\usage{
modplot_samplescheck(
  age.seq,
  post,
  input.list,
  staninputdf,
  simdata,
  minage = 0
)
}
\description{
Model to Sample Comparison plot
}
