% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{p_link_dataset}
\alias{p_link_dataset}
\title{Link Posterior estimates for a dataset into output}
\usage{
p_link_dataset(ageseq, post, input.list, dataset.i, dataset.size = NULL)
}
\description{
Works for both 1 and many species
}
