% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{p_link_species}
\alias{p_link_species}
\title{Link Posterior estimates for a species into output}
\usage{
p_link_species(ageseq, post, input.list, species.i, dataset.size = NULL)
}
\description{
Works for both 1 and many species
}
