% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{plot_modtosample}
\alias{plot_modtosample}
\title{Model to Sample Comparison plot}
\usage{
plot_modtosample(
  age.seq,
  post,
  input.list,
  dataset.is = NULL,
  minages = NULL,
  names.key = NULL,
  return.data = FALSE
)
}
\description{
Model to Sample Comparison plot
}
