% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulate static survival sample}
\usage{
run_simulation(
  sim.in = NULL,
  include_r = TRUE,
  include_e = TRUE,
  include_s = TRUE,
  sample_size = 200,
  age.maturity = 0,
  show.plot = TRUE,
  max.age = 150
)
}
\description{
Simulate static survival sample
}
