% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_activepars}
\alias{get_activepars}
\title{Get list of parameters actually active in the model}
\usage{
get_activepars(out, mod, input.list)
}
\description{
Get list of parameters actually active in the model
}
