% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{plot_ageX}
\alias{plot_ageX}
\title{Age X plot}
\usage{
plot_ageX(
  X,
  post,
  age.seq,
  species.is,
  minages = NULL,
  names.key = NULL,
  return.data = FALSE
)
}
\description{
Age X plot
}
