% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputplotting_functions.R
\name{plot_posteriorsurvival}
\alias{plot_posteriorsurvival}
\title{Model predicted survival curves}
\usage{
plot_posteriorsurvival(
  post,
  age.seq,
  input.list,
  names.key = NULL,
  sim.in = NULL,
  N = 150
)
}
\description{
Model predicted survival curves
}
